调整无归传送方式
